requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
fastapi-cors